pyproject.toml
[project]
name = "https-secutity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mitmproxy>=12.2.2",
    "python-dotenv>=1.2.2",
]
